requirements.txt
flask
werkzeug
gunicorn
mysql-connector-python
python-dotenv
